Remove the test for getpublishtimeAgo function because its not pure Change the name parameter for this function Relatse #35
